+++ remote/grpc_client_test.go
@@ -1,4 +1,4 @@
-package rpc
+package remote
 
 import (
 	"bytes"
