+++ remote/mdb_service.go
@@ -29,10 +29,10 @@ type service struct {
 	cacheMutex sync.RWMutex
 }
 
-// StartService wraps an existing Jasper manager in a mongo wire protocol
+// StartMDBService wraps an existing Jasper manager in a MongoDB wire protocol
 // service and starts it. The caller is responsible for closing the connection
 // using the returned jasper.CloseFunc.
-func StartService(ctx context.Context, m jasper.Manager, addr net.Addr) (util.CloseFunc, error) { //nolint: interfacer
+func StartMDBService(ctx context.Context, m jasper.Manager, addr net.Addr) (util.CloseFunc, error) { //nolint: interfacer
 	host, p, err := net.SplitHostPort(addr.String())
 	if err != nil {
 		return nil, errors.Wrap(err, "invalid address")
